Move test directories
