+++ jest.config.js
@@ -20,8 +20,8 @@ module.exports = {
   // framework before each test
   moduleNameMapper: {
     '\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$':
-      '<rootDir>/__mocks__/fileMock.js',
-    '\\.(scss|css)$': '<rootDir>/__mocks__/styleMock.js',
+      '<rootDir>/tests/mocks/fileMock.js',
+    '\\.(scss|css)$': '<rootDir>/tests/mocks/styleMock.js',
   },
   setupFilesAfterEnv: ['./jest.setup.js'],
 };
