+++ tests/components/DashCount.test.js
