+++ tests/components/DashCountGroup.test.js
