+++ tests/fixtures/bz-issue-counts.js
