+++ tests/lib/const.test.js
