+++ tests/lib/utils/projects.test.js
