+++ tests/mocks/fileMock.js
