+++ tests/mocks/styleMock.js
