+++ tests/pages/api/gh-team.test.js
@@ -1,25 +1,26 @@
 import fetchMock from 'fetch-mock';
 import MockExpressRequest from 'mock-express-request';
 import MockExpressResponse from 'mock-express-response';
-import getIssueCounts from 'pages/api/gh-issue-counts';
-import issueCountData from 'fixtures/gh-issue-counts';
+import getTeam from 'pages/api/gh-team';
+import teamData from 'tests/fixtures/gh-team';
 
 describe(__filename, () => {
   beforeEach(() => {
-    fetchMock.mock('https://api.github.com/graphql', issueCountData);
+    fetchMock.mock('https://api.github.com/graphql', teamData);
   });
 
   afterEach(() => {
     fetchMock.restore();
+    fetchMock.reset();
   });
 
-  it('should return issue count data', async () => {
+  it('should return team data', async () => {
     const req = new MockExpressRequest({
       method: 'GET',
-      url: '/api/gh-issue-counts/',
+      url: '/api/gh-team/',
     });
     const res = new MockExpressResponse();
-    await getIssueCounts(req, res);
-    expect(res._getJSON()).toEqual(issueCountData);
+    await getTeam(req, res);
+    expect(res._getJSON()).toEqual(teamData);
   });
 });
