+++ tests/pages/contrib/contrib-welcome.test.js
@@ -1,15 +1,15 @@
 import fetchMock from 'fetch-mock';
 import * as nextRouter from 'next/router';
+import ghContribWelcomeData from 'tests/fixtures/gh-contrib-welcome';
 import { cleanup, render } from '@testing-library/react';
-import ghGoodFirstBugsData from 'fixtures/gh-good-first-bugs';
-import GoodFirstBugs, {
+import ContribWelcome, {
   getServerSideProps,
-} from 'pages/contrib/good-first-bugs';
+} from 'pages/contrib/contrib-welcome';
 
 describe(__filename, () => {
   beforeEach(() => {
     nextRouter.useRouter = jest.fn();
-    fetchMock.mock(/\/api\/gh-good-first-bugs\//, ghGoodFirstBugsData);
+    fetchMock.mock(/\/api\/gh-contrib-welcome\//, ghContribWelcomeData);
   });
 
   afterEach(() => {
@@ -17,16 +17,16 @@ describe(__filename, () => {
     cleanup();
   });
 
-  it('should render the Good First Bugs Page', async () => {
+  it('should render the Contrib Welcome Page', async () => {
     nextRouter.useRouter.mockImplementation(() => ({
-      pathname: '/contrib/good-first-bugs/',
+      pathname: '/contrib/contrib-welcome/',
       query: {
         dir: 'asc',
         sort: 'updatedAt',
       },
     }));
     const { props } = await getServerSideProps();
-    const { findByRole } = render(<GoodFirstBugs {...props} />);
+    const { findByRole } = render(<ContribWelcome {...props} />);
     const main = await findByRole('main');
     expect(main).toHaveClass('container');
   });
@@ -34,7 +34,7 @@ describe(__filename, () => {
   it('should fetch data via getServerSideProps', async () => {
     const { props: serverProps } = await getServerSideProps();
     expect(
-      serverProps.goodFirstBugsData.data.good_first_bugs.results.length,
+      serverProps.contribWelcomeData.data.contrib_welcome.results.length,
     ).toEqual(3);
   });
 });
