+++ tests/pages/milestones/[milestone].test.js
@@ -1,21 +1,18 @@
 import fetchMock from 'fetch-mock';
 import * as nextRouter from 'next/router';
-import ghProjectsData from 'fixtures/gh-projects';
-import ghTeamData from 'fixtures/gh-team';
+import ghMilestoneIssuesData from 'tests/fixtures/gh-milestone-issues';
 import { cleanup, render } from '@testing-library/react';
-import Projects, { getServerSideProps } from 'pages/projects/[year]/[quarter]';
+import Milestones, { getServerSideProps } from 'pages/milestones/[milestone]';
 
 describe(__filename, () => {
   let fakeProps;
 
   beforeEach(() => {
     nextRouter.useRouter = jest.fn();
-    fetchMock.mock(/\/api\/gh-projects\//, ghProjectsData);
-    fetchMock.mock(/\/api\/gh-team\//, ghTeamData);
+    fetchMock.mock(/\/api\/gh-milestone-issues\//, ghMilestoneIssuesData);
     fakeProps = {
       params: {
-        year: '2021',
-        quarter: 'Q1',
+        milestone: '2021-01-21',
       },
     };
   });
@@ -25,24 +22,23 @@ describe(__filename, () => {
     cleanup();
   });
 
-  it('should render the Projects Page', async () => {
+  it('should render the Milestone Page', async () => {
     nextRouter.useRouter.mockImplementation(() => ({
-      pathname: '/projects/2021/Q1/',
+      pathname: '/milestones/2021-01-21/',
       query: {
-        year: '2021',
-        quarter: 'Q1',
+        milestone: '2021-01-21',
+        dir: 'asc',
+        sort: 'assignee',
       },
     }));
     const { props } = await getServerSideProps(fakeProps);
-    const { findByRole } = render(<Projects {...props} />);
+    const { findByRole } = render(<Milestones {...props} />);
     const main = await findByRole('main');
     expect(main).toHaveClass('container');
   });
 
   it('should fetch data via getServerSideProps', async () => {
     const { props: serverProps } = await getServerSideProps(fakeProps);
-    expect(
-      serverProps.projects.data.organization.projects.nodes.length,
-    ).toEqual(4);
+    expect(serverProps.milestoneIssues.length).toEqual(13);
   });
 });
