+++ tests/pages/projects/[year]/[quarter].test.js
@@ -1,18 +1,21 @@
 import fetchMock from 'fetch-mock';
 import * as nextRouter from 'next/router';
-import ghMilestoneIssuesData from 'fixtures/gh-milestone-issues';
+import ghProjectsData from 'tests/fixtures/gh-projects';
+import ghTeamData from 'tests/fixtures/gh-team';
 import { cleanup, render } from '@testing-library/react';
-import Milestones, { getServerSideProps } from 'pages/milestones/[milestone]';
+import Projects, { getServerSideProps } from 'pages/projects/[year]/[quarter]';
 
 describe(__filename, () => {
   let fakeProps;
 
   beforeEach(() => {
     nextRouter.useRouter = jest.fn();
-    fetchMock.mock(/\/api\/gh-milestone-issues\//, ghMilestoneIssuesData);
+    fetchMock.mock(/\/api\/gh-projects\//, ghProjectsData);
+    fetchMock.mock(/\/api\/gh-team\//, ghTeamData);
     fakeProps = {
       params: {
-        milestone: '2021-01-21',
+        year: '2021',
+        quarter: 'Q1',
       },
     };
   });
@@ -22,23 +25,24 @@ describe(__filename, () => {
     cleanup();
   });
 
-  it('should render the Milestone Page', async () => {
+  it('should render the Projects Page', async () => {
     nextRouter.useRouter.mockImplementation(() => ({
-      pathname: '/milestones/2021-01-21/',
+      pathname: '/projects/2021/Q1/',
       query: {
-        milestone: '2021-01-21',
-        dir: 'asc',
-        sort: 'assignee',
+        year: '2021',
+        quarter: 'Q1',
       },
     }));
     const { props } = await getServerSideProps(fakeProps);
-    const { findByRole } = render(<Milestones {...props} />);
+    const { findByRole } = render(<Projects {...props} />);
     const main = await findByRole('main');
     expect(main).toHaveClass('container');
   });
 
   it('should fetch data via getServerSideProps', async () => {
     const { props: serverProps } = await getServerSideProps(fakeProps);
-    expect(serverProps.milestoneIssues.length).toEqual(13);
+    expect(
+      serverProps.projects.data.organization.projects.nodes.length,
+    ).toEqual(4);
   });
 });
